+++ airquality-pir/mqtt_handler.h
@@ -9,3 +9,6 @@ void mqtt_init(Client& wifi_client, std::function<void (char *, uint8_t *, unsig
 void mqtt_connect(const char* uuid);
 
 void mqtt_notify_value(const char* device_uuid, const char* feature_uuid, const char* type, float value);
+
+void mqtt_notify_alarm(const char* device_uuid, const char* feature_uuid, const char* feature_name,
+                       const char* alarm_type, float value);